pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safety-compliance-manager"
version = "0.1.0"
description = "AI-powered Safety & Compliance Management agent for Microsoft Office 365"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "anthropic>=0.39.0",
    "sqlalchemy[asyncio]>=2.0.23",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "python-pptx>=0.6.23",
    "msal>=1.25.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]

[project.scripts]
scm = "src.cli.main:app"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
